+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Version 1.31.1 - 2024-12-02
+
+### Fixed
+
+- Fixed level track message for unlocking Wizard mode at level 268. (Thanks Katie O'Connell for reporting!)
+
 ## Version 1.31.0 - 2024-12-02
 
 ### Added
