+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "block-sorting",
-  "version": "1.31.0",
+  "version": "1.31.1",
   "description": "Mobile Block sorting game, as a Progressive Web App",
   "author": {
     "name": "Matthijs Groen",
